# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manage OAuth-connected Gmail and Microsoft mailboxes, proxies, and synced messages.",
        "title": "DeezMails API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/accounts/{id}/sync": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Queue a mailbox sync job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Folder ID or all",
                        "name": "folder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.JobRun"
                        }
                    }
                }
            }
        },
        "/api/jobs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List durable job runs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter account",
                        "name": "accountId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.JobRun"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Clear completed and failed job history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/jobs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a durable job run",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.JobRun"
                        }
                    }
                }
            }
        },
        "/api/jobs/{id}/retry": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Retry a failed job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.JobRun"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Account": {
            "type": "object",
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "folder": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incomingHost": {
                    "type": "string"
                },
                "incomingPort": {
                    "type": "integer"
                },
                "lastCheckedAt": {
                    "type": "string"
                },
                "lastError": {
                    "type": "string"
                },
                "lastErrorAction": {
                    "$ref": "#/definitions/models.MailErrorAction"
                },
                "lastErrorCode": {
                    "$ref": "#/definitions/models.MailErrorCode"
                },
                "lastSyncedAt": {
                    "type": "string"
                },
                "nextSyncAt": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "proxy": {
                    "$ref": "#/definitions/models.Proxy"
                },
                "proxyId": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "syncEnabled": {
                    "type": "boolean"
                },
                "tlsMode": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.JobRun": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/models.Account"
                },
                "accountId": {
                    "type": "integer"
                },
                "attempts": {
                    "type": "integer"
                },
                "created": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "endedAt": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "errorAction": {
                    "$ref": "#/definitions/models.MailErrorAction"
                },
                "errorCode": {
                    "$ref": "#/definitions/models.MailErrorCode"
                },
                "fetched": {
                    "type": "integer"
                },
                "folder": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logs": {
                    "type": "string"
                },
                "nextAttemptAt": {
                    "type": "string"
                },
                "queueId": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.JobStatus"
                },
                "synced": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.JobType"
                },
                "updated": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "queued",
                "running",
                "retrying",
                "succeeded",
                "failed"
            ],
            "x-enum-varnames": [
                "JobStatusQueued",
                "JobStatusRunning",
                "JobStatusRetrying",
                "JobStatusSucceeded",
                "JobStatusFailed"
            ]
        },
...
        "models.JobType": {
            "type": "string",
            "enum": [
                "sync_account",
                "verify_connection"
            ],
            "x-enum-varnames": [
                "JobTypeSyncAccount",
                "JobTypeVerifyConnection"
            ]
        },
        "models.MailErrorAction": {
            "type": "string",
            "enum": [
                "none",
                "retry",
                "reconnect",
                "edit_account",
                "contact_admin"
            ],
            "x-enum-varnames": [
                "MailErrorActionNone",
                "MailErrorActionRetry",
                "MailErrorActionReconnect",
                "MailErrorActionEditAccount",
                "MailErrorActionContactAdmin"
            ]
        },
        "models.MailErrorCode": {
            "type": "string",
            "enum": [
                "authentication_failed",
                "permission_denied",
                "organization_policy",
                "quota_exceeded",
                "rate_limited",
                "provider_unavailable",
                "not_found",
                "invalid_configuration",
                "tls_failed",
                "network_failed",
                "protocol_failed",
                "response_too_large",
                "cancelled",
                "internal_error"
            ],
            "x-enum-varnames": [
                "MailErrorAuthentication",
                "MailErrorPermission",
                "MailErrorPolicy",
                "MailErrorQuota",
                "MailErrorRateLimit",
                "MailErrorUnavailable",
                "MailErrorNotFound",
                "MailErrorConfiguration",
                "MailErrorTLS",
                "MailErrorNetwork",
                "MailErrorProtocol",
                "MailErrorResponseTooLarge",
                "MailErrorCancelled",
                "MailErrorInternal"
            ]
        },
        "models.Proxy": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}